move residual to networks
