+++ bayesflow/experimental/resnet/__init__.py
@@ -1,3 +1,2 @@
 from .dense_resnet import DenseResNet
 from .resnet import ResNet
-from .residual import Residual
